Prevent essentials home replace
Prevent essentials home replace with players with old player data. This also prevents home replacement sometimes with players who player before but the server is still busy fetching the data. This happened multiple times in my server, when a server went down and they got moved back to it, they all got their homes replaced because the server had a wrong response on fetching data, teleporting them as if they were new players. This will force-check if the player already has a home.